requirements.txt
absl-py==0.7.1
astor==0.7.1
astroid==2.1.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
entrypoints==0.3
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.6
numpy==1.16.2
pandocfilters==1.4.2
parso==0.3.4
pbr==5.1.3
pickleshare==0.7.5
Pillow==5.4.1
Pillow-PIL==0.1.dev0
pip==19.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.0
Pygments==2.3.1
pylint==2.2.2
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pywinpty==0.5.5
pyzmq==18.0.1
qtconsole==4.4.3
scipy==1.2.1
Send2Trash==1.5.0
setuptools==39.0.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
typed-ast==1.2.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.1
wheel==0.33.1
widgetsnbextension==3.4.2
wrapt==1.11.1
